Added BibTeX file for pCT Collaboration References
References to literature used in our pCT publications is continually
increasing and references to newly generated papers are not distributed
to other collaborators in a systematic way.  Newly created LaTeX
directories for new papers typically results in a .bib file being copied
from a previous paper and if it is not copied from the most recent, any
additions made while writing that paper do not make it into the new .bib
file.  By creating a main pct.bib file which everyone contributes to and
uses, we can propagate references to our newly created papers and papers
we have cited to the rest of the collaboration.  Each time a new paper
is to be written, the newest version of pct.bib can be pulled from
GitHub and additions made can be merged so we each have access to the
most current set of references to the collaboration's publications as
well as their referenced works.  This will also improve the way we cite
each other's work.