Added pCT Data Naming/Organizational Format
Documents showing hierarchical structure of pCT data storage and
file/directory naming scheme.  This format is shown visually in 1 PDF
and an explanation of each of the files/folders shown is given in
another PDF.  Note that the current version of the visual representation
of the format has the hierarchy under "Experimental" cut off, but its
subdirectories should match those under the GEANT4 and TOPAS hierarchies
exactly except for the date folder at the top of the hierarchy does not
have an "G_" or "T_" identifier and is just formatted as "MMDDYYYY".
Everything else matches exactly.