Updated format and fixed PDF diagram
format updates:
- Changed date format from MMDDYYYY to YY-MM-DD so lexicographic
ordering organizes by date with year first and adding dashes between
parts makes reading/interpreting dates easier
- Changed run # format from "Run_XXX" to "XXX[_AAA]" removing the
unnecessary inclusion of the word "Run" which is also not included in
the file naming of the raw data and also adding the optional inclusion
of a descriptor tag "_AAA" which is included in the run # portion of the
raw data file names in some cases, such as in the case of the head
phantom where "_Inf" or "_Sup" is added to differentiate between the
scans of the inferior and superior regions.  This new format then
matches the format used in the logging and naming of the raw data files
as part of the effort to establish consistency in the format used by
those generating the data and the collaboration wide format used in
cases where this is possible.
- Added definitions of the format and organizational scheme used for the
separate storage of each type of data: raw, preprocessed, projection,
and reconstruction data.  The purpose of this approach is to provide a
separate directory to store the data generated by different people and
allow them to store this data in a format that is convenient for them as
long as this format is consistent and provides the object/phantom name
and creation dates of the input/output data. The format for the raw data
has been confirmed and the preprocessed data follows this general scheme
as well, though this has not been confirmed yet.  The projection data,
which is stored separately from the preprocessed data although the
contents of the preprocessed and projection data is the same, since the
existing file naming scheme follows that of the raw data and the
projection data files generated are simply links to these files, they
are stored separately in essentially the same format as the raw data to
minimize the difference between the storage format of this data on the
local computers of those that generate this data and its shared storage
so existing analysis programs etc can still be used with minimal
changes.  The projection data links generated from these files are then
organized with the projection data files generated directly by Pier's
preprocessing program so the reconstruction program can read input data
from this directory in the same way regardless of whether the projection
data files are links or the actual data itself.  With each type of data
stored separately, the organized data directory
(/pCT_data/organized_data)  will simply contain links to this data as
opposed to an additional copy of the data itself.  The data stored on
Akek was almost completely organized according to the collaboration wide
format and the organized data directory contained the actual data
itself, not links to the data since the various types of data were not
stored/organized separately.  In migrating the data from Akek to Kodiak,
most of the raw, preprocessed, projection, and reconstruction data was
separated into their individual directories but in an effort to preserve
the existing organization, the data was duplicated from the
organized_data directory instead of moved from it.  Long term, the
organized_data directory will contain little to no actual data just soft
links to the data, and the duplicated data currently stored in this
directory will gradually be replaced by soft links to the data from the
individual directories containing the actual data.