Updated/modified phantom ID and object folder naming schemes and data distribution users
Finished the Excel file that lists the phantom ID used to name raw data
files for each phantom currently in possession of the collaboration and
the corresponding object folder name to use for storing/organizing the
data for each of these phantoms.  This file also outlines the additions
to the naming scheme used for non standard scan configurations such as
superior/inferior scans of the head phantom and for continuous scans,
which are referred to as subclasses and added to the file names as tags
after the run #.  This standardization provides the basis for a script
that parses raw data files to determine the phantom ID, run #, angle,
and any subclass tags to create/organize links to this data.  The
script, uses the phantom ID to lookup the corresponding object folder
name of the directory where the data should be placed, uses the run #
and subclass tags to determine the appropriate name for the run #
folder, then creates any necessary directories and creates/organizes
soft links to this data.  By creating/organizing soft links to this
data, the raw data can be stored by scan date, which is convenient for
analysis of the beam/scanner for a particular scan date, while the links
are stored by object/dates which is important as the directory hierarchy
establishes the data dependencies at each stage of reconstruction.

This commit also contains a CAD drawing of the pCT data organizational
hierarchy, the list of approved data distribution users, and renamed
several other files so their file names more closely linked to what
their contents are.