Modified pCT experimental file naming file
After 07/25/15 run, some changes were made to the way files will be
named during the scan, although the naming scheme used did not conform
to the format established prior to the run.  These are minor changes
made in an attempt to come closer to the naming scheme that was used at
the time of the run which is clearly what feels more natural when
creating them.  These are minor changes.